Move cachedfile -> cache
